backend/requirements.txt
fastapi
uvicorn
websockets
google-genai
python-dotenv
pydantic
pydantic-settings
